User service now uses email service instead of nodemailer. A few other small improvements to user service. Completed password reset tests.

+++ api/users/user.service.js
@@ -8,7 +8,7 @@ const jwt = require('jsonwebtoken');
 const bcrypt = require('bcryptjs');
 const db = require('../db');
 const { v4: uuidv4 } = require('uuid');
-const nodemailer = require("nodemailer");
+const emailService = require("../email.service");
 const User = db.User;
 var _settings;
 
@@ -90,43 +90,35 @@ async function update(id, userParam) {
 }
 
 async function _delete(id) {
+    if (await User.count() === 1) {
+        throw "Unable to delete last user";
+    }
+    const user = await User.findById(id);
+    if (!user) {
+        throw "User not found";
+    }
     await User.findByIdAndRemove(id);
 }
 
-async function generateResetToken(req, username) {
+async function generateResetToken(req, username) { // TODO: decouple from http (req)
     const user = await User.findOne({ "username": username });
     checkValidUser(user);
 
     // generate reset token and save in user
     let token = uuidv4();
-    update(user.id, { resetToken: token });
+    await update(user.id, { resetToken: token });
 
-    // setup email server
-    let transporter = nodemailer.createTransport({
-        host: config.smtp.host, // TODO: make all smtp settings web based
-        port: config.smtp.port,
-        secure: config.smtp.ssl, // true for 465, false for other ports
-        auth: { user: config.smtp.user, pass: config.smtp.password }
-    });
-
+    // build email message
     let rootPath = "/ui";
     if (typeof settings().ui !== "undefined" && typeof settings().ui.path !== "undefined") {
         rootPath = settings().ui.path.length ? "/" + settings().ui.path : "";
     }
-
     let resetLink = req.protocol + "://" + req.get('host') + rootPath + "/#/authentication/reset-password?t=" + token;
-    let resetVerbiage = "A password reset has been request for username '" + user.username + 
-                        "'. Please click this link to change your password: " + resetLink;
+    let message = "A password reset has been request for username '" + user.username + 
+                "'. Please click this link to change your password: " + resetLink;
 
-    // send email
-    let info = await transporter.sendMail({
-        from: config.smtp.fromAddress,
-        to: user.email,
-        subject: "Password Reset for " + user.username,
-        text: resetVerbiage,
-        html: resetVerbiage
-    });
-    return true;
+    await emailService.send(user.email, "Password Reset for " + user.username, message);
+    return token;
 }
 
 async function resetPassword(token, { password }) { // TODO: need web page to handle new password input
@@ -136,7 +128,7 @@ async function resetPassword(token, { password }) { // TODO: need web page to ha
     if (!user) throw "Invalid or expired reset token";
 
     // update user password
-    return update(user.id, { password: password, resetToken: null });
+    return await update(user.id, { password: password, resetToken: null });
 }
 
 function checkValidUser(user) {
@@ -179,7 +171,12 @@ function checkStrongPassword(password) {
 
 function settings() {
     if (!_settings) {
-        _settings = require(process.env.RED_SETTINGS_FILE);
+        try {
+            _settings = require(process.env.RED_SETTINGS_FILE);
+        }
+        catch (e) {
+            return {};
+        }
     }
     return _settings;
 }
